adjust code and revert changes to completions
