+++ crates/tinymist-query/src/analysis/bib.rs
@@ -1,7 +1,7 @@
 use typst::{
     foundations::{Bytes, IntoValue, Packed},
     introspection::Introspector,
-    model::{BibliographyElem, CslStyle},
+    model::BibliographyElem,
 };
 use yaml_rust2::{parser::Event, parser::MarkedEventReceiver, scanner::Marker};
 
@@ -214,58 +214,6 @@ impl YamlBib {
         let entries = loader.content.into_iter().filter_map(to_entry).collect();
         Self { entries }
     }
-}
-
-/// Render the citation string in the bib with given CSL style.
-/// Returns a pair of (citation, bib_item).
-pub(crate) fn render_citation_string(
-    bib_info: &BibInfo,
-    style: &CslStyle,
-    key: &str,
-    support_html: bool,
-) -> Option<(String, String)> {
-    use hayagriva::{
-        BibliographyDriver, BibliographyRequest, BufWriteFormat, CitationItem, CitationRequest,
-        ElemChildren,
-    };
-
-    let entry = bib_info.entries.get(key)?;
-    let raw_entry = entry.raw_entry.as_ref()?;
-
-    let mut driver = BibliographyDriver::new();
-
-    let locales = &[];
-    driver.citation(CitationRequest::from_items(
-        vec![CitationItem::with_entry(raw_entry)],
-        style.get(),
-        locales,
-    ));
-
-    let result = driver.finish(BibliographyRequest {
-        style: style.get(),
-        locale: None, // todo: get locale from CiteElem
-        locale_files: locales,
-    });
-    let rendered_bib = result.bibliography?;
-
-    let format_elem = |elem: &ElemChildren| {
-        let mut buf = String::new();
-        elem.write_buf(
-            &mut buf,
-            if support_html {
-                BufWriteFormat::Html
-            } else {
-                BufWriteFormat::Plain
-            },
-        )
-        .ok()?;
-        Some(buf)
-    };
-
-    Some((
-        format_elem(&result.citations.first()?.citation)?,
-        format_elem(&rendered_bib.items.first()?.content)?,
-    ))
 }
 
 #[cfg(test)]
