+++ crates/tinymist-query/src/lib.rs
@@ -55,6 +55,7 @@ mod adt;
 mod lsp_typst_boundary;
 mod prelude;
 
+mod bib;
 mod code_action;
 mod code_context;
 mod code_lens;
